Fixed issues with dash and ultimate
There was a bug where you could use the ultimate when still on the start screen, so I fixed it. There was also a bug where if you died and the dash or ultimate cooldown was not finished, it would finish on the start screen and you could still use those abilities, so I fixed that too.