Changed controller script
Changed a lot of controller stuff, such as how moving the cursor is calculated. Also made it so if you turn the controller on, it only locks the cursor to the window if you are using the controller and not your mouse. Still need to fix some thing with cursor movement.